+++ src/test/java/BasicMAPF/TestUtils.java
@@ -113,7 +113,7 @@ public class TestUtils {
         }
 
         @Override
-        public float getHToTargetFromLocation(I_Coordinate target, I_Location currLocation) {
+        public int getHToTargetFromLocation(I_Coordinate target, I_Location currLocation) {
             return 0;
         }
